Move protocol extension to TurboNavigatingCoordinator
